--- app/frontend/src/views/designer/Root.vue
@@ -1,18 +1,0 @@
-<template>
-  <FormDesigner :formId="formId" :formVersionId="formVersionId" />
-</template>
-
-<script>
-import FormDesigner from '@/components/designer/FormDesigner.vue';
-
-export default {
-  name: 'Editor',
-  components: {
-    FormDesigner
-  },
-  props: {
-    formId: String,
-    formVersionId: String
-  }
-};
-</script>
