+++ app/frontend/src/views/form/Teams.vue
@@ -1,5 +1,6 @@
 <template>
   <div>
+    <v-breadcrumbs :items="breadcrumbs" />
     <h1 class="my-6 text-center">Team Management</h1>
     <TeamManagement :formId="f" />
   </div>
@@ -11,13 +12,31 @@ import TeamManagement from '@/components/forms/TeamManagement.vue';
 export default {
   name: 'FormTeams',
   components: {
-    TeamManagement
+    TeamManagement,
   },
   props: {
     f: {
       type: String,
-      required: true
-    }
-  }
+      required: true,
+    },
+  },
+  computed: {
+    breadcrumbs() {
+      const path = [
+        {
+          text: 'Form',
+        },
+        {
+          text: 'Manage Form'
+        }
+      ];
+      if (this.$route.meta.breadcrumbTitle) {
+        path.push({
+          text: this.$route.meta.breadcrumbTitle,
+        });
+      }
+      return path;
+    },
+  },
 };
 </script>
