+++ app/frontend/src/views/user/Forms.vue
@@ -1,6 +1,6 @@
 <template>
   <div>
-    <h1 class="text-center">My Forms</h1>
+    <h1 class="my-6 text-center">My Forms</h1>
     <div class="text-right">
       <strong>
         <router-link :to="{ name: 'FormDesigner' }">Create a New Form</router-link>
